requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.2
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.5.1
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-google==10.26.0
apache-airflow-providers-http==4.13.3
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.1
apache-airflow-providers-sqlite==3.9.1
apispec==6.7.1
argcomplete==3.5.1
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==43.0.3
db-dtypes==1.3.1
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.9
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.10.0
gcloud-aio-auth==5.3.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.3.0
gcsfs==2024.10.0
google-ads==25.1.0
google-analytics-admin==0.23.2
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.73.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-automl==2.14.1
google-cloud-batch==0.17.31
google-cloud-bigquery==3.27.0
google-cloud-bigquery-datatransfer==3.17.1
google-cloud-bigtable==2.27.0
google-cloud-build==3.27.1
google-cloud-compute==1.22.0
google-cloud-container==2.54.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.23.0
google-cloud-dataflow-client==0.8.14
google-cloud-dataform==0.5.13
google-cloud-dataplex==2.4.0
google-cloud-dataproc==5.15.1
google-cloud-dataproc-metastore==1.16.1
google-cloud-dlp==3.25.1
google-cloud-kms==3.1.1
google-cloud-language==2.15.1
google-cloud-logging==3.11.3
google-cloud-memcache==1.10.1
google-cloud-monitoring==2.23.1
google-cloud-orchestration-airflow==1.15.1
google-cloud-os-login==2.15.1
google-cloud-pubsub==2.27.1
google-cloud-redis==2.16.1
google-cloud-resource-manager==1.13.1
google-cloud-run==0.10.13
google-cloud-secret-manager==2.21.1
google-cloud-spanner==3.50.1
google-cloud-speech==2.28.1
google-cloud-storage==2.18.2
google-cloud-storage-transfer==1.13.1
google-cloud-tasks==2.17.1
google-cloud-texttospeech==2.21.1
google-cloud-translate==3.18.0
google-cloud-videointelligence==2.14.1
google-cloud-vision==3.8.1
google-cloud-workflows==1.15.1
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.68.0
grpcio-gcp==0.2.2
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
json-merge-patch==0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==24.20.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
ordered-set==4.1.0
packaging==24.2
pandas==2.1.4
pandas-gbq==0.24.0
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
py4j==0.10.9.7
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydata-google-auth==1.9.0
Pygments==2.18.0
PyJWT==2.10.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyspark==3.5.3
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.21.0
rsa==4.9
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools==75.6.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.12.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.7.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.2
statsd==4.0.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.17.0
WTForms==3.2.1
yarl==1.18.0
zipp==3.21.0
